# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aigoni API
  version: 1.0.0
  description: |
    Aigoni API contract. `x-implementation-status` marks routes registered now
    implemented REST contract.
servers:
  - url: /
...
paths:
...
  /api/admin/v1/settings:
    get:
      tags: [Admin API]
      summary: Get safe site settings
      x-implementation-status: implemented
      security: [{AdminSession: []}]
      responses:
        '200':
          description: Settings without secrets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
    patch:
      tags: [Admin API]
      summary: Update site settings
      x-implementation-status: implemented
      security: [{AdminSession: [], CSRFToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsPatch'
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '422': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: aigoni_session
      description: HttpOnly, SameSite=Lax; Secure over HTTPS.
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on Admin write operations in the target contract.
    APIKeyHeader: {type: apiKey, in: header, name: X-API-Key}
    BearerAuth: {type: http, scheme: bearer}
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  schemas:
...
    Settings:
      type: object
      properties:
        site:
          type: object
          properties: {name: {type: string}, description: {type: string}, author: {type: string}, base_url: {type: string, format: uri}, logo: {type: string}, author_avatar: {type: string}, utc_offset: {type: string}}
        pagination:
          type: object
          properties: {posts_per_page: {type: integer}, home_posts_count: {type: integer}}
        paths:
          type: object
          properties: {content_dir: {type: string}, posts_dir: {type: string}, pages_dir: {type: string}, notes_dir: {type: string}, public_dir: {type: string}, uploads_dir: {type: string}}
        uploads:
          type: object
          properties: {allowed_extensions: {type: array, items: {type: string}}, max_bytes: {type: integer}, site_asset_path: {type: string}}
        updated_at: {type: string, format: date-time}
    SettingsPatch:
      type: object
      properties:
        site:
          type: object
          properties: {name: {type: string}, description: {type: string}, author: {type: string}, base_url: {type: string, format: uri}, logo: {type: string}, author_avatar: {type: string}, utc_offset: {type: string}}
        pagination:
          type: object
          properties: {posts_per_page: {type: integer}, home_posts_count: {type: integer}}
        paths:
          type: object
          properties: {content_dir: {type: string}, posts_dir: {type: string}, pages_dir: {type: string}, notes_dir: {type: string}, public_dir: {type: string}, uploads_dir: {type: string}}
...
    ErrorResponse: {type: object, required: [error], properties: {error: {oneOf: [{type: string}, {$ref: '#/components/schemas/ErrorDetail'}]}}}
    ErrorDetail: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}, field_errors: {type: object, additionalProperties: {type: string}}, request_id: {type: string}}}